Moving beam transform to separate header file
